Add actual CLI binary and slow tests to hit it
