+++ Gemfile
@@ -4,3 +4,4 @@ gem 'sorcerer', "~> 0.3.0"
 gem "rspec", "~> 2.10.0"
 gem "rake", "~> 0.9.2.2"
 gem "cucumber", "~> 1.2"
+gem "aruba", "~> 0.4"
