+++ bin/raffle
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby
+$: << File.dirname(__FILE__) + '/../lib'
+require 'raffle/cli'
+require 'raffle/file_system'
+
+file_system = Raffle::FileSystem.new
+Raffle::CLI.new(file_system).run(*ARGV)
