+++ lib/raffle/file_system.rb
@@ -1,0 +1,13 @@
+module Raffle
+  class FileSystem
+    def read(path)
+      File.read(path)
+    end
+
+    def write(path, content)
+      File.open(path, 'w') do |file|
+        file.puts content
+      end
+    end
+  end
+end
